feat!: use extractor pattern for request handlers
This is a massive change and the transition probably isn't even fully
done yet. This should be a good starting point, though.

The remaining work includes:

* Removing more of the `RequestExt` trait
* Creating `IntoResponse` trait so that the consumers can return other
  types than `cot::Result<Response>`
* More work on maintaining consistency in the API

Note that this is a (quite big) breaking change, mostly because it
removes a lot of methods from the `RequestExt` trait.

This change is needed to implement automatic OpenAPI spec generation
in #159.